added migrations and context db

+++ MoviesApi.Data.Models/Rating.cs
@@ -5,7 +5,7 @@ namespace MoviesApi.Data.Models;
 public class Rating
 {
     [Key]
-    public long MovieId { get; set; }
+    public int MovieId { get; set; }
     public double RatingValue { get; set; }
     public long Votes { get; set; }
